Add tools for checking that errors don't accidentally change
Adds tools/gen_expected_errors.c which is used to create tests/data/expected_errors.txt. This is then used in test_d2itemreader_errors to test that the errors are still the same when d2itemreader_parse_any is run against the entire corpus of fuzzed inputs.

This might be a bit brittle and not particularly helpful in the future, but the basic idea was very useful when checking that the iterative parsing refactor behaved the same as the previous implementation so I figured I'd formalize it/integrate it with the test suite to ensure that the error behavior doesn't change accidentally.